chore: remove pundit, use native code
